+++ srcs/pipe_utils.c
@@ -6,7 +6,7 @@
 /*   By: lpassera <lpassera@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/19 10:25:57 by pcharton          #+#    #+#             */
-/*   Updated: 2021/07/28 17:05:37 by lpassera         ###   ########.fr       */
+/*   Updated: 2021/08/04 14:24:29 by lpassera         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,8 +26,8 @@ t_pipe	*init_pipeline_utils(t_pipeline *pipeline)
 	t->index = -1;
 	t->pipe_tab = allocate_pipe_tab(pipeline->pipe_count - 1);
 	g_globals->pids = malloc(sizeof(int) * pipeline->pipe_count - 1);
-	if (!t->pipe_tab || !g_globals->pids) // WARNING IT COULD LEAK HERE
-		return (NULL);
+	if (!t->pipe_tab || !g_globals->pids)
+		return (free_pipeline_utils(t, pipeline));
 	pipe_node = pipeline->commands;
 	while (pipe_node)
 	{
@@ -41,7 +41,7 @@ t_pipe	*init_pipeline_utils(t_pipeline *pipeline)
 void	clean_up_pipeline_utils(t_pipe *tmp, t_pipeline *pipeline)
 {
 	wait_pipeline_end(pipeline->pipe_count - 1);
-	deallocate_pipe_tab(tmp->pipe_tab, pipeline->pipe_count - 1);
+	deallocate_pipe_tab(tmp->pipe_tab, pipeline->pipe_count - 1, true);
 	restore_in_and_out(&tmp->in_and_out);
 	free(g_globals->pids);
 	free(tmp);
@@ -63,30 +63,13 @@ int	**allocate_pipe_tab(int	nb)
 			tab[i] = malloc(sizeof(int) * 2);
 			if (!tab[i])
 			{
-				deallocate_pipe_tab(tab, i);
+				deallocate_pipe_tab(tab, i, false);
 				return (NULL);
 			}
 		}
 		tab[i] = NULL;
 	}
 	return (tab);
-}
-
-void	deallocate_pipe_tab(int **tab, int nb)
-{
-	int	i;
-
-	i = 0;
-	while (i < nb)
-	{
-		close(tab[i][0]);
-		close(tab[i][1]);
-		free(tab[i]);
-		tab[i] = NULL;
-		i++;
-	}
-	free(tab);
-	tab = NULL;
 }
 
 int	wait_pipeline_end(int pipe_count)
